requirements.txt
Flask==2.3.2
twilio==8.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
